split react component and moved file
